move rake code to its own file
